Change MaxOutputTokens and NumPredict types from int to int64 for better scalability

+++ yomawari/generative-ai/abstractions/chatcompletion/chatoptions.go
@@ -5,7 +5,7 @@ import "github.com/futugyou/yomawari/generative-ai/abstractions"
 type ChatOptions struct {
 	ChatThreadId         *string                `json:"chatThreadId,omitempty"`
 	Temperature          *float64               `json:"temperature,omitempty"`
-	MaxOutputTokens      *int                   `json:"maxOutputTokens,omitempty"`
+	MaxOutputTokens      *int64                 `json:"maxOutputTokens,omitempty"`
 	TopP                 *float64               `json:"topP,omitempty"`
 	TopK                 *int                   `json:"topK,omitempty"`
 	FrequencyPenalty     *float64               `json:"frequencyPenalty,omitempty"`
